Part of Guy Harris' Jan-93 rewrite, including: Add in support for multiple levels of continuation.
Add in support for "beshort", "belong", "bedate", "leshort",
"lelong", and "ledate" types.

Fix the old-style ANDing - don't change it to a MASK op
relation, because:

	there's already code to handle old-style ANDing and there's no
	need to use the MASK op code to implement this;

	the parser that used to be there assumed that there was a
	*second* number that represented a value against which
	to compare the ANDed value, and the old-style ANDing
	*didn't* have that so lines containing the old-style
	syntax "0 byte &0x80 dynamically linked" turned into
	lines containing the new-style syntax "0 byte&0x80 0
	dynamically linked", as the missing second value
	defaulted to 0.

That may be where that bizarro "!074000,000000" stuff from the
Clipper magic number stuff came, but since you can just
*explicitly* implement that with new-style ANDing, one might as
well just do that....

Also, make old-style "not set"ting - "0 byte ^0x80 statically
linked", meaning "this line matches if any of the bits in the
value *aren't* set in the number from the file" - work.