Comment #3 From Jakub Jelinek  	 on 2006-10-06 07:42 EST  	[reply] The file is the stripped-into-separate-file debuginfo, in which sections that are present in the stripped file are made SHT_NOBITS and not present at all. So, file needs to check for SHT_NOBITS .note sections and don't look at them at all.  Currently it just reads some completely unrelated data (probably part of one of the .debug_* sections).
Comment #6 From Jakub Jelinek 	on 2006-10-17 10:48 EST 	[reply]
Proposed patch.  While it would be possible to do what I proposed (i.e.
for each PT_NOTE look up in section header table in which section the virtual
address falls in and test if it is not SHT_NOBITS), this is far easier and
faster
and has a nice side-effect that it doesn't duplication information in file
output.  Unpatched file will e.g. say:
file /bin/bash
/bin/bash: ELF 64-bit LSB executable, AMD x86-64, version 1 (SYSV), for
GNU/Linux 2.6.9, dynamically linked (uses shared libs), for GNU/Linux 2.6.9,
stripped
(note the .note.ABI-tag section being parsed and printed twice, once when found
through PT_NOTE, once when found through SHT_NOTE).  When the executable has
a section header table, then it is safe to assume that either there will be
a SHT_NOTE section corresponding to the PT_NOTE program header, or it will be
SHT_NOBITS in a debuginfo file (at which point it is not present).